# Conflicts:
#	src/Umbraco.Web.BackOffice/Mapping/ContentMapDefinition.cs
#	src/Umbraco.Web.BackOffice/Security/BackOfficeAntiforgery.cs
#	src/Umbraco.Web.UI.Client/src/common/directives/components/property/umbproperty.directive.js

+++ src/Umbraco.Infrastructure/Logging/Viewer/LogViewerConfig.cs
@@ -1,45 +1,51 @@
-﻿using System.Collections.Generic;
-using System.Linq;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Persistence.Repositories;
 using Umbraco.Cms.Core.Scoping;
+using IScope = Umbraco.Cms.Infrastructure.Scoping.IScope;
 
-namespace Umbraco.Cms.Core.Logging.Viewer
+namespace Umbraco.Cms.Core.Logging.Viewer;
+
+public class LogViewerConfig : ILogViewerConfig
 {
-    public class LogViewerConfig : ILogViewerConfig
+    private readonly ILogViewerQueryRepository _logViewerQueryRepository;
+    private readonly IScopeProvider _scopeProvider;
+
+    public LogViewerConfig(ILogViewerQueryRepository logViewerQueryRepository, IScopeProvider scopeProvider)
     {
-        private readonly ILogViewerQueryRepository _logViewerQueryRepository;
-        private readonly IScopeProvider _scopeProvider;
+        _logViewerQueryRepository = logViewerQueryRepository;
+        _scopeProvider = scopeProvider;
+    }
 
-        public LogViewerConfig(ILogViewerQueryRepository logViewerQueryRepository, IScopeProvider scopeProvider)
+    public IReadOnlyList<SavedLogSearch> GetSavedSearches()
+    {
+        using IScope scope = _scopeProvider.CreateScope(autoComplete: true);
+        IEnumerable<ILogViewerQuery> logViewerQueries = _logViewerQueryRepository.GetMany();
+        SavedLogSearch[] result = logViewerQueries.Select(x => new SavedLogSearch() { Name = x.Name, Query = x.Query }).ToArray();
+        return result;
+    }
+
+    public IReadOnlyList<SavedLogSearch> AddSavedSearch(string name, string query)
+    {
+        using IScope scope = _scopeProvider.CreateScope(autoComplete: true);
+        _logViewerQueryRepository.Save(new LogViewerQuery(name, query));
+
+        return GetSavedSearches();
+    }
+
+    [Obsolete("Use the overload that only takes a 'name' parameter instead. This will be removed in Umbraco 14.")]
+    public IReadOnlyList<SavedLogSearch> DeleteSavedSearch(string name, string query) => DeleteSavedSearch(name);
+
+    public IReadOnlyList<SavedLogSearch> DeleteSavedSearch(string name)
+    {
+        using IScope scope = _scopeProvider.CreateScope(autoComplete: true);
+        ILogViewerQuery? item = _logViewerQueryRepository.GetByName(name);
+
+        if (item is not null)
         {
-            _logViewerQueryRepository = logViewerQueryRepository;
-            _scopeProvider = scopeProvider;
-        }
-
-        public IReadOnlyList<SavedLogSearch> GetSavedSearches()
-        {
-            using var scope = _scopeProvider.CreateScope(autoComplete: true);
-            var logViewerQueries =  _logViewerQueryRepository.GetMany();
-            var result = logViewerQueries.Select(x => new SavedLogSearch() { Name = x.Name, Query = x.Query }).ToArray();
-            return result;
-        }
-
-        public IReadOnlyList<SavedLogSearch> AddSavedSearch(string name, string query)
-        {
-            using var scope = _scopeProvider.CreateScope(autoComplete: true);
-            _logViewerQueryRepository.Save(new LogViewerQuery(name, query));
-
-            return GetSavedSearches();
-        }
-
-        public IReadOnlyList<SavedLogSearch> DeleteSavedSearch(string name, string query)
-        {
-            using var scope = _scopeProvider.CreateScope(autoComplete: true);
-            var item = _logViewerQueryRepository.GetByName(name);
             _logViewerQueryRepository.Delete(item);
-            //Return the updated object - so we can instantly reset the entire array from the API response
-            return GetSavedSearches();
         }
+
+        // Return the updated object - so we can instantly reset the entire array from the API response
+        return GetSavedSearches();
     }
 }
